Add tabs and change style
