--- assets/language/ranok_words.txt
@@ -1,22 +1,0 @@
-arg
-return
-var
-const
-abs
-sqrt
-sin
-cos
-tan
-arctan
-arcsin
-arccos
-cosh
-sinh
-tanh
-exp
-ln
-log
-log10
-log2
-ceil
-floor
