+++ res.qrc
@@ -16,12 +16,14 @@
         <file>assets/styles/NeonButtons.qss</file>
         <file>assets/styles/QTDark.qss</file>
         <file>assets/styles/Ubuntu.qss</file>
-        <file>assets/language/ranok_words.txt</file>
         <file>shaders/voxel.frag</file>
         <file>shaders/voxel.geom</file>
         <file>shaders/voxel.vert</file>
         <file>shaders/grid.frag</file>
         <file>shaders/grid.vert</file>
         <file>assets/images/saveIcon.png</file>
+        <file>assets/images/Code.svg</file>
+        <file>assets/images/Cube1.svg</file>
+        <file>assets/images/Cube2.svg</file>
     </qresource>
 </RCC>
