+++ src/Gui/CodeEditor/CodeEditor.cpp
@@ -18,23 +18,18 @@ void CodeEditor::AddFile(const QString &sourceFile)
     QFile file(sourceFile);
     if(file.open(QIODevice::ReadOnly))
     {
-        auto tab = new CodeEditorTab(this);
-//        connect(data.tab, &CodeEditorTab::textChanged, this, &CodeEditor::tabTextChanged);
-        tab->setPlainText(file.readAll());
-        tab->verticalScrollBar()->setSliderPosition(0);
-
-//        data.filepath = sourceFile;
         auto splitFilepath = sourceFile.split('/');
         if(splitFilepath.empty())
             splitFilepath = sourceFile.split('\\');
-//        data.filename = splitFilepath.last();
+        QString fileName = splitFilepath.last();
 
-
-        addTab(tab, "tab");
-        setCurrentIndex(addTab(tab, "tab"));
-
-
+        auto tab = new CodeEditorTab(this);
+        tab->setPlainText(file.readAll());
+        tab->verticalScrollBar()->setSliderPosition(0);
         m_tabs.push_back(tab);
+
+        int newTabId = addTab(tab, fileName);
+        setCurrentIndex(newTabId);
     }
     else
         QMessageBox::critical(this, "Ошибка", "Невозможно открыть файл");
