+++ src/ScreensView.h
@@ -1,0 +1,21 @@
+#ifndef SCREENSVIEW_H
+#define SCREENSVIEW_H
+
+#include <QTabWidget>
+
+class ScreensView : public QTabWidget
+{
+    Q_OBJECT
+public:
+    explicit ScreensView(QWidget *parent = nullptr);
+
+
+signals:
+
+
+private:
+
+
+};
+
+#endif // SCREENSVIEW_H
